move fetching intents to lib directory #26
